server/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docker==7.0.0
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
packaging==23.2
requests==2.31.0
sortedcontainers==2.4.0
urllib3==2.1.0
Werkzeug==3.0.1
aiohttp==3.3.2
flask==2.2.2
flask-cors==3.0.10
pandas==1.5.3
numpy==1.24.1
mysql-connector-python==8.0.32